requirements.txt
Flask
requests
spotipy
spotipy
uuid
numpy
pandas
gunicorn
lyricsgenius 
flask_sslify
mysql-connector-python
sklearn